+++ backend/db/postgres/modules/c3-taxonomy/module.sql
@@ -5,3 +5,4 @@
 \ir ../../schema/20_capability_coverage_views.sql
 \ir ../../schema/25_capability_governance.sql
 \ir ../../schema/34_c3_board_state.sql
+\ir ../../schema/38_c3_entity_link_view.sql